Fix Task name resolution
- Tasks from the modern providers are ints, not longs. So are
  Opcodes. Update the model to reflect this.
- Tasks must be resolvable from the modern provider even when the
  description comes from the legacy provider.

+++ src/EventLogExpert.Library/EventProviderDatabase/EventProviderDbContext.cs
@@ -48,10 +48,10 @@ public class EventProviderDbContext : DbContext
 
         modelBuilder.Entity<ProviderDetails>()
             .Property(e => e.Opcodes)
-            .HasConversion<JsonValueConverter<Dictionary<long, string>>>();
+            .HasConversion<JsonValueConverter<Dictionary<int, string>>>();
 
         modelBuilder.Entity<ProviderDetails>()
             .Property(e => e.Tasks)
-            .HasConversion<JsonValueConverter<Dictionary<long, string>>>();
+            .HasConversion<JsonValueConverter<Dictionary<int, string>>>();
     }
 }
